Catch `E325: ATTENTION`.
If `E325: ATTENTION` is thrown during `exec` (in case the user tries to open
a file that has a swap file already exists for), Vim will present a dialog
to let the user decide what to do. If the user selects Quit from this
dialog, an error processing `PickCommand()` will be reported unless we catch
this `E325: ATTENTION`.