Allow `pick` commands to unescape the selection
This commit adds the behavior of allowing `pick` commands to opt into
escaping the output. As commands such as `:tag` does not want to escape
its identifier, whilst commands such as `:edit` wants to.

Before this commit, when you were trying to use `PickTag` to select a
method such as `paid?`. The following error would be raised:

```viml
Error detected while processing function PickTag[1]..PickCommand:
E426: tag not found: paid\?
```

This commit changes `PickTag()` not to escape the selection, but to pass
it along as is.